lighter trace command to increase the reusability and testability of the trace methods
